% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{as_id}
\alias{as_id}
\title{Safely coerce to ID character vector}
\usage{
as_id(x)
}
\arguments{
\item{x}{Vector to coerce}
}
\value{
Character vector
}
\description{
Safely coerce to ID character vector
}
\keyword{internal}
