% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{assert_columns}
\alias{assert_columns}
\title{Validate that a data frame has required columns}
\usage{
assert_columns(data, required, context = "check")
}
\arguments{
\item{data}{Data frame}

\item{required}{Character vector of required column names}

\item{context}{Character describing the calling context}
}
\description{
Validate that a data frame has required columns
}
\keyword{internal}
