% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{assert_numeric}
\alias{assert_numeric}
\title{Validate that a column is numeric}
\usage{
assert_numeric(data, col, context = "check")
}
\arguments{
\item{data}{Data frame}

\item{col}{Column name}

\item{context}{Calling context}
}
\description{
Validate that a column is numeric
}
\keyword{internal}
