% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_result.R
\name{bind_check_results}
\alias{bind_check_results}
\title{Combine multiple check results into a summary tibble}
\usage{
bind_check_results(...)
}
\arguments{
\item{...}{check_result objects or lists of check_result objects}
}
\value{
A tibble summarizing all checks
}
\description{
Combine multiple check results into a summary tibble
}
