% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_missing.R
\name{check_all_missing_variables}
\alias{check_all_missing_variables}
\title{Check for variables with 100\% missing values}
\usage{
check_all_missing_variables(data, ...)
}
\arguments{
\item{data}{Data frame}

\item{...}{Reserved}
}
\value{
A check_result object
}
\description{
C.04: Variables with zero non-missing values.
}
