% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_timing.R
\name{check_collection_window}
\alias{check_collection_window}
\title{Check collection window}
\usage{
check_collection_window(data, id_col, date_col, start_date, end_date, ...)
}
\arguments{
\item{data}{Data frame of survey submissions}

\item{id_col}{Character. Name of the primary key column}

\item{date_col}{Character. Name of the column containing collection dates}

\item{start_date}{Character or Date. Start of the valid collection window}

\item{end_date}{Character or Date. End of the valid collection window}

\item{...}{Reserved for future use}
}
\value{
A check_result object
}
\description{
F.02: Flag surveys collected outside the expected date window.
}
