% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_metadata.R
\name{check_consent}
\alias{check_consent}
\title{Check consent status}
\usage{
check_consent(data, id_col, consent_col, consent_value = 1, ...)
}
\arguments{
\item{data}{Data frame of survey submissions}

\item{id_col}{Character. Name of the primary key column}

\item{consent_col}{Character. Name of the consent column}

\item{consent_value}{Value indicating consent was given (default 1)}

\item{...}{Reserved for future use}
}
\value{
A check_result object
}
\description{
A.06: Flag records where consent was not given.
}
