% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_paradata.R
\name{check_device_consistency}
\alias{check_device_consistency}
\title{Check device consistency per enumerator}
\usage{
check_device_consistency(data, id_col, enum_col, device_col, ...)
}
\arguments{
\item{data}{Data frame of survey submissions}

\item{id_col}{Character. Name of the primary key column}

\item{enum_col}{Character. Name of the enumerator ID column}

\item{device_col}{Character. Name of the device identifier column}

\item{...}{Reserved for future use}
}
\value{
A check_result object
}
\description{
N.08: Flag enumerators who used more than one device during data collection.
Groups by enumerator and counts distinct device identifiers. Enumerators
with more than one device are flagged.
}
