% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_enumerator.R
\name{check_dk_rate}
\alias{check_dk_rate}
\title{Check don't-know response rate by enumerator}
\usage{
check_dk_rate(
  data,
  id_col,
  enum_col,
  dk_value = "dk",
  check_cols = NULL,
  threshold_sd = 2,
  ...
)
}
\arguments{
\item{data}{Data frame of survey submissions}

\item{id_col}{Character. Name of the primary key column}

\item{enum_col}{Character. Name of the enumerator ID column}

\item{dk_value}{Character. The "don't know" string value (default "dk")}

\item{check_cols}{Character vector. Columns to inspect; if NULL, all columns
except id_col and enum_col are used}

\item{threshold_sd}{Numeric. Number of SDs above team mean to trigger flag (default 2)}

\item{...}{Reserved for future use}
}
\value{
A check_result object
}
\description{
B.03: Flag enumerators with an unusually high rate of "don't know"
responses (more than threshold_sd SDs above team mean).
}
\details{
A cell is counted as DK if it matches dk_value (case-insensitive) or
equals -99, -88, or -77.
}
