% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_duplicates.R
\name{check_duplicate_ids}
\alias{check_duplicate_ids}
\title{Check for duplicate survey IDs}
\usage{
check_duplicate_ids(data, id_col, ...)
}
\arguments{
\item{data}{Data frame of survey submissions}

\item{id_col}{Character. Name of the primary key column}

\item{...}{Reserved for future use}
}
\value{
A check_result object
}
\description{
A.01: Finds records sharing the same primary key.
}
