% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_logic.R
\name{check_duration_by_hh_size}
\alias{check_duration_by_hh_size}
\title{Check survey duration by household size}
\usage{
check_duration_by_hh_size(
  data,
  id_col,
  duration_col,
  hh_size_col,
  min_minutes_per_member = 3,
  ...
)
}
\arguments{
\item{data}{Data frame of survey submissions}

\item{id_col}{Character. Name of the primary key column}

\item{duration_col}{Character. Name of the duration column (in minutes)}

\item{hh_size_col}{Character. Name of the household size column}

\item{min_minutes_per_member}{Numeric. Minimum expected minutes per HH member
(default 3)}

\item{...}{Reserved for future use}
}
\value{
A check_result object
}
\description{
F.06: Flag surveys that are shorter than expected given the household size.
}
