% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_enumerator.R
\name{check_enumerator_duration}
\alias{check_enumerator_duration}
\title{Check enumerator survey duration}
\usage{
check_enumerator_duration(
  data,
  id_col,
  enum_col,
  duration_col,
  threshold_sd = 2,
  ...
)
}
\arguments{
\item{data}{Data frame of survey submissions}

\item{id_col}{Character. Name of the primary key column}

\item{enum_col}{Character. Name of the enumerator ID column}

\item{duration_col}{Character. Name of the survey duration column (numeric, in minutes)}

\item{threshold_sd}{Numeric. Number of SDs below team mean to trigger flag (default 2)}

\item{...}{Reserved for future use}
}
\value{
A check_result object
}
\description{
B.02: Flag enumerators whose mean survey duration is suspiciously low
(more than threshold_sd standard deviations below the team mean).
}
