% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_enumerator.R
\name{check_enumerator_time_gap}
\alias{check_enumerator_time_gap}
\title{Check time gaps between consecutive surveys by enumerator}
\usage{
check_enumerator_time_gap(
  data,
  id_col,
  enum_col,
  start_time_col,
  min_gap_minutes = 5,
  ...
)
}
\arguments{
\item{data}{Data frame of survey submissions}

\item{id_col}{Character. Name of the primary key column}

\item{enum_col}{Character. Name of the enumerator ID column}

\item{start_time_col}{Character. Name of the survey start time column
(parsed with as.POSIXct())}

\item{min_gap_minutes}{Numeric. Minimum expected gap in minutes (default 5)}

\item{...}{Reserved for future use}
}
\value{
A check_result object
}
\description{
B.08: Flag surveys where the gap from the previous survey start time
is impossibly short (less than min_gap_minutes).
}
