% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_paradata.R
\name{check_field_comments}
\alias{check_field_comments}
\title{Check field comments for supervisor review}
\usage{
check_field_comments(data, id_col, comment_col, ...)
}
\arguments{
\item{data}{Data frame of survey submissions}

\item{id_col}{Character. Name of the primary key column}

\item{comment_col}{Character. Name of the comment column}

\item{...}{Reserved for future use}
}
\value{
A check_result object
}
\description{
N.04: Extract and list all non-empty field comments. This is informational:
comments need human review and are flagged for supervisor attention.
}
