% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_metadata.R
\name{check_form_version}
\alias{check_form_version}
\title{Check form version}
\usage{
check_form_version(data, id_col, version_col, expected_version, ...)
}
\arguments{
\item{data}{Data frame of survey submissions}

\item{id_col}{Character. Name of the primary key column}

\item{version_col}{Character. Name of the form version column}

\item{expected_version}{Character or numeric. The expected (current) form version}

\item{...}{Reserved for future use}
}
\value{
A check_result object
}
\description{
A.07: Flag records using an outdated form version.
}
