% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_timing.R
\name{check_future_dates}
\alias{check_future_dates}
\title{Check for future dates}
\usage{
check_future_dates(data, id_col, date_col, reference_date = Sys.Date(), ...)
}
\arguments{
\item{data}{Data frame of survey submissions}

\item{id_col}{Character. Name of the primary key column}

\item{date_col}{Character. Name of the column containing dates}

\item{reference_date}{Date. Reference date to compare against (default Sys.Date())}

\item{...}{Reserved for future use}
}
\value{
A check_result object
}
\description{
F.03: Flag records with dates in the future relative to a reference date.
}
