% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_gps.R
\name{check_gps_accuracy}
\alias{check_gps_accuracy}
\title{Check GPS accuracy}
\usage{
check_gps_accuracy(data, id_col, accuracy_col, max_accuracy = 50, ...)
}
\arguments{
\item{data}{Data frame of survey submissions}

\item{id_col}{Character. Name of the primary key column}

\item{accuracy_col}{Character. Name of the GPS accuracy column (metres)}

\item{max_accuracy}{Numeric. Maximum acceptable accuracy in metres (default 50)}

\item{...}{Reserved for future use}
}
\value{
A check_result object
}
\description{
E.03: Flag GPS readings with accuracy greater than a threshold.
}
