% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_gps.R
\name{check_gps_null_island}
\alias{check_gps_null_island}
\title{Check GPS null island and default values}
\usage{
check_gps_null_island(data, id_col, lat_col, lon_col, ...)
}
\arguments{
\item{data}{Data frame of survey submissions}

\item{id_col}{Character. Name of the primary key column}

\item{lat_col}{Character. Name of the latitude column}

\item{lon_col}{Character. Name of the longitude column}

\item{...}{Reserved for future use}
}
\value{
A check_result object
}
\description{
E.07: Flag GPS coordinates at (0, 0), or with lat/lon equal to known
default or placeholder values (0, 999, -999, 9999).
}
