% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_outliers.R
\name{check_hard_range}
\alias{check_hard_range}
\title{Check values against hard range constraints}
\usage{
check_hard_range(data, id_col, num_col, min_val = -Inf, max_val = Inf, ...)
}
\arguments{
\item{data}{Data frame of survey submissions}

\item{id_col}{Character. Name of the primary key column}

\item{num_col}{Character. Name of the numeric column to check}

\item{min_val}{Numeric. Minimum allowed value (default -Inf)}

\item{max_val}{Numeric. Maximum allowed value (default Inf)}

\item{...}{Reserved for future use}
}
\value{
A check_result object
}
\description{
D.04: Flag values outside a fixed [min_val, max_val] range.
}
