% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_metadata.R
\name{check_id_format}
\alias{check_id_format}
\title{Check ID format against expected pattern}
\usage{
check_id_format(data, id_col, pattern, ...)
}
\arguments{
\item{data}{Data frame of survey submissions}

\item{id_col}{Character. Name of the primary key column}

\item{pattern}{Character. Regular expression pattern that valid IDs must match}

\item{...}{Reserved for future use}
}
\value{
A check_result object
}
\description{
A.04: Flag IDs that do not match the expected regex pattern.
}
