% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_duplicates.R
\name{check_id_in_sample}
\alias{check_id_in_sample}
\title{Check IDs against sampling frame}
\usage{
check_id_in_sample(data, id_col, sampling_frame, ...)
}
\arguments{
\item{data}{Data frame}

\item{id_col}{Primary key column}

\item{sampling_frame}{Character vector of valid IDs from master list}

\item{...}{Reserved}
}
\value{
A check_result object
}
\description{
A.05: IDs not in the master sampling list.
}
