% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_logic.R
\name{check_income_expenditure}
\alias{check_income_expenditure}
\title{Check income-expenditure ratio}
\usage{
check_income_expenditure(
  data,
  id_col,
  income_col,
  expenditure_col,
  max_ratio = 5,
  ...
)
}
\arguments{
\item{data}{Data frame of survey submissions}

\item{id_col}{Character. Name of the primary key column}

\item{income_col}{Character. Name of the income column}

\item{expenditure_col}{Character. Name of the expenditure column}

\item{max_ratio}{Numeric. Maximum acceptable expenditure/income ratio (default 5)}

\item{...}{Reserved for future use}
}
\value{
A check_result object
}
\description{
G.04: Flag records where expenditure exceeds a plausible ratio of income.
}
