% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_missing.R
\name{check_missing_by_enumerator}
\alias{check_missing_by_enumerator}
\title{Check missing rate by enumerator}
\usage{
check_missing_by_enumerator(data, id_col, enum_col, threshold_sd = 2, ...)
}
\arguments{
\item{data}{Data frame}

\item{id_col}{Primary key column}

\item{enum_col}{Enumerator ID column}

\item{threshold_sd}{Numeric. Number of SD above team mean to flag (default 2)}

\item{...}{Reserved}
}
\value{
A check_result object
}
\description{
C.02: Flag enumerators with unusually high missing rates.
}
