% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_duplicates.R
\name{check_missing_ids}
\alias{check_missing_ids}
\title{Check for missing primary IDs}
\usage{
check_missing_ids(data, id_col, ...)
}
\arguments{
\item{data}{Data frame}

\item{id_col}{Primary key column}

\item{...}{Reserved}
}
\value{
A check_result object
}
\description{
A.03: Records with NA or empty ID field.
}
