% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_text_quality.R
\name{check_name_format}
\alias{check_name_format}
\title{Check name format validity}
\usage{
check_name_format(data, id_col, name_col, ...)
}
\arguments{
\item{data}{Data frame of survey submissions}

\item{id_col}{Character. Name of the primary key column}

\item{name_col}{Character. Name of the name column to check}

\item{...}{Reserved for future use}
}
\value{
A check_result object
}
\description{
O.04: Flags names that are NA/empty, only digits, known test values,
or single characters. Useful for catching enumerator shortcuts and
data entry mistakes in respondent name fields.
}
