% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_outliers.R
\name{check_negative_values}
\alias{check_negative_values}
\title{Check for negative values}
\usage{
check_negative_values(data, id_col, num_col, ...)
}
\arguments{
\item{data}{Data frame of survey submissions}

\item{id_col}{Character. Name of the primary key column}

\item{num_col}{Character. Name of the numeric column to check}

\item{...}{Reserved for future use}
}
\value{
A check_result object
}
\description{
D.09: Flag negative values in columns that should be non-negative
(e.g., counts, areas, quantities).
}
