% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_text_quality.R
\name{check_other_specify}
\alias{check_other_specify}
\title{Check "other (specify)" write-in responses}
\usage{
check_other_specify(data, id_col, other_col, ...)
}
\arguments{
\item{data}{Data frame of survey submissions}

\item{id_col}{Character. Name of the primary key column}

\item{other_col}{Character. Name of the "other specify" text column}

\item{...}{Reserved for future use}
}
\value{
A check_result object
}
\description{
O.01: Extracts all unique "other (specify)" write-in responses for review.
Flags records that have non-empty other_col values so they can be recoded
or validated during data cleaning.
}
