% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_outliers.R
\name{check_outliers_iqr}
\alias{check_outliers_iqr}
\title{Check outliers using IQR method}
\usage{
check_outliers_iqr(data, id_col, num_col, multiplier = 1.5, ...)
}
\arguments{
\item{data}{Data frame of survey submissions}

\item{id_col}{Character. Name of the primary key column}

\item{num_col}{Character. Name of the numeric column to check}

\item{multiplier}{Numeric. IQR multiplier for fence calculation (default 1.5)}

\item{...}{Reserved for future use}
}
\value{
A check_result object
}
\description{
D.01: Flag values below Q1 - multiplier\emph{IQR or above Q3 + multiplier}IQR.
}
