% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_paradata.R
\name{check_photo_attachment}
\alias{check_photo_attachment}
\title{Check photo attachment completeness}
\usage{
check_photo_attachment(data, id_col, photo_col, ...)
}
\arguments{
\item{data}{Data frame of survey submissions}

\item{id_col}{Character. Name of the primary key column}

\item{photo_col}{Character. Name of the photo attachment column}

\item{...}{Reserved for future use}
}
\value{
A check_result object
}
\description{
N.06: Flag records where the photo column is missing, empty, or contains
a placeholder value such as "none".
}
