% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_logic.R
\name{check_roster_completeness}
\alias{check_roster_completeness}
\title{Check roster completeness}
\usage{
check_roster_completeness(data, id_col, hh_size_col, member_count_col, ...)
}
\arguments{
\item{data}{Data frame of survey submissions}

\item{id_col}{Character. Name of the primary key column}

\item{hh_size_col}{Character. Name of the reported household size column}

\item{member_count_col}{Character. Name of the actual roster member count column}

\item{...}{Reserved for future use}
}
\value{
A check_result object
}
\description{
C.07: Flag records where roster member count does not match the
reported household size.
}
