% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_logic.R
\name{check_survey_end_before_start}
\alias{check_survey_end_before_start}
\title{Check survey end time before start time}
\usage{
check_survey_end_before_start(data, id_col, start_col, end_col, ...)
}
\arguments{
\item{data}{Data frame of survey submissions}

\item{id_col}{Character. Name of the primary key column}

\item{start_col}{Character. Name of the survey start time column}

\item{end_col}{Character. Name of the survey end time column}

\item{...}{Reserved for future use}
}
\value{
A check_result object
}
\description{
F.04: Flag surveys where the end timestamp is before the start timestamp.
}
