% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_fabrication.R
\name{check_variance_ratio}
\alias{check_variance_ratio}
\title{Check variance ratio by enumerator}
\usage{
check_variance_ratio(data, id_col, enum_col, num_cols, max_ratio = 2, ...)
}
\arguments{
\item{data}{Data frame of survey submissions}

\item{id_col}{Character. Name of the primary key column}

\item{enum_col}{Character. Name of the enumerator ID column}

\item{num_cols}{Character vector. Numeric columns to test}

\item{max_ratio}{Numeric. Maximum acceptable between/within variance ratio
(default 2)}

\item{...}{Reserved for future use}
}
\value{
A check_result object
}
\description{
M.08: For each numeric variable, compute the ratio of between-enumerator
variance to within-enumerator variance. A high ratio suggests enumerator
identity explains too much of the variability, possibly indicating
fabrication or systematic bias.
}
