% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_results.R
\name{.build_sheet_corrections}
\alias{.build_sheet_corrections}
\title{Build Sheet: Corrections Log}
\usage{
.build_sheet_corrections(ctx)
}
\description{
Build Sheet: Corrections Log
}
\keyword{internal}
