% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_results.R
\name{.build_sheet_skip_patterns}
\alias{.build_sheet_skip_patterns}
\title{Build Sheet: Skip Violations}
\usage{
.build_sheet_skip_patterns(ctx)
}
\description{
Build Sheet: Skip Violations
}
\keyword{internal}
