% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_results.R
\docType{data}
\name{.check_descriptions}
\alias{.check_descriptions}
\title{Human-readable descriptions for common check codes}
\format{
An object of class \code{character} of length 28.
}
\usage{
.check_descriptions
}
\description{
Human-readable descriptions for common check codes
}
\keyword{internal}
