% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_results.R
\name{.extract_results}
\alias{.extract_results}
\title{Extract a flat list of check_result objects from a report or raw list}
\usage{
.extract_results(report)
}
\arguments{
\item{report}{An adc_report or list of check_result objects}
}
\value{
List of check_result objects
}
\description{
Extract a flat list of check_result objects from a report or raw list
}
\keyword{internal}
