% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_results.R
\docType{data}
\name{.sheet_registry}
\alias{.sheet_registry}
\title{Registry of all available sheets}
\format{
An object of class \code{list} of length 18.
}
\usage{
.sheet_registry
}
\description{
Registry of all available sheets
}
\keyword{internal}
