% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_gps.R
\name{haversine_km}
\alias{haversine_km}
\title{Compute Haversine distance between two points}
\usage{
haversine_km(lat1, lon1, lat2, lon2)
}
\arguments{
\item{lat1}{Latitude of point 1 (degrees)}

\item{lon1}{Longitude of point 1 (degrees)}

\item{lat2}{Latitude of point 2 (degrees)}

\item{lon2}{Longitude of point 2 (degrees)}
}
\value{
Distance in kilometres
}
\description{
Compute Haversine distance between two points
}
\keyword{internal}
