% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_result.R
\name{is_check_result}
\alias{is_check_result}
\title{Test if an object is a check_result}
\usage{
is_check_result(x)
}
\arguments{
\item{x}{Object to test}
}
\value{
Logical
}
\description{
Test if an object is a check_result
}
