% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_data.R
\name{platform_defaults}
\alias{platform_defaults}
\title{Get platform-specific default column mappings}
\usage{
platform_defaults(platform, cols)
}
\arguments{
\item{platform}{Character. Platform name}

\item{cols}{Character vector. Available column names in the data}
}
\value{
Named list of default column mappings
}
\description{
Get platform-specific default column mappings
}
\keyword{internal}
