% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_checks.R
\name{print.adc_report}
\alias{print.adc_report}
\title{Print an adc_report summary}
\usage{
\method{print}{adc_report}(x, ...)
}
\arguments{
\item{x}{An adc_report object}

\item{...}{Ignored}
}
\description{
Print an adc_report summary
}
