% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_result.R
\name{print.check_result}
\alias{print.check_result}
\title{Print a check result}
\usage{
\method{print}{check_result}(x, ...)
}
\arguments{
\item{x}{A check_result object}

\item{...}{Ignored}
}
\description{
Print a check result
}
