% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{read_config}
\alias{read_config}
\title{Read and validate a project configuration file}
\usage{
read_config(path)
}
\arguments{
\item{path}{Path to YAML config file}
}
\value{
A validated config list of class "adc_config"
}
\description{
Read and validate a project configuration file
}
