% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_checks.R
\name{summary.adc_report}
\alias{summary.adc_report}
\title{Summarize an adc_report}
\usage{
\method{summary}{adc_report}(object, ...)
}
\arguments{
\item{object}{An adc_report object}

\item{...}{Ignored}
}
\value{
A tibble summarizing all checks
}
\description{
Returns the summary tibble produced by \code{\link[=bind_check_results]{bind_check_results()}}.
}
