% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{validate_config}
\alias{validate_config}
\title{Validate config structure}
\usage{
validate_config(cfg)
}
\arguments{
\item{cfg}{List from YAML}
}
\description{
Validate config structure
}
\keyword{internal}
