% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_results.R
\name{.build_corrections_template}
\alias{.build_corrections_template}
\title{Build the Corrections template sheet}
\usage{
.build_corrections_template(flagged_tbl)
}
\description{
Build the Corrections template sheet
}
\keyword{internal}
